Organizes folder, commands into separate files
